+++ packages/svg/src/index.ts
@@ -1,10 +1,10 @@
+export { svgExtension } from "./extension/svg-extension";
+export { SVGMeasure } from "./measurement/svg-measure";
 export {
-	svgExtension,
-	SVGMeasure,
 	type SvgDimensions,
 	type SvgElement,
 	type SvgNode,
 	type SvgToPdfColor,
 	type SvgToPdfFontOptions,
 	type SvgToPdfOptions,
-} from "./svg-measure";
+} from "./types";